move to utils_global
